Refactor world structure: consolidate World2d and World3d into a single World class, updating relevant includes in SceneNode and Tilemap2d.
